update method Application.java and converters

+++ src/main/java/ru/levelup/homework5/dto/Temperature.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class Temperature {
 
-    private Long inputTemperature;
+    private double inputTemperature;
 }
